Price and Quantity editing is added to the new frontend

+++ Tour-of-Flags/src/app/shopping-cart/shopping-cart.component.ts
@@ -1,7 +1,7 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 
 import { Account } from '../account';
-import { ShoppingCart } from '../shopping-cart';
+import { CartItem } from '../cart-item';
 import { AccountService } from '../account.service';
 import { Flag } from '../flag';
 import { FlagService } from '../flag.service';
@@ -11,15 +11,13 @@ import { FlagService } from '../flag.service';
   templateUrl: './shopping-cart.component.html',
   styleUrls: ['./shopping-cart.component.css']
 })
-export class ShoppingCartComponent {
-  currentAccount!: Account;
-  cart: ShoppingCart | undefined;
+export class ShoppingCartComponent implements OnInit {
+  currentAccount?: Account;
+  cart: CartItem[] = [];
 
   constructor(private accountService: AccountService, private flagService: FlagService) {
-    if (accountService.getCurrentAccount()) {
-      this.currentAccount = accountService.getCurrentAccount()!;
-    }
-   }
+    this.currentAccount = accountService.getCurrentAccount();
+  }
 
   ngOnInit(): void {
     this.getCart();
@@ -28,25 +26,29 @@ export class ShoppingCartComponent {
   getCart(): void {
     if (this.currentAccount) {
       this.accountService.getCart(this.currentAccount.username)
-      .subscribe(cart => this.cart = cart);
+        .subscribe(cart => this.cart = cart ?? []);
     }
   }
 
-  getFlagFromID(id: number): any {
-    // Need to figure this out
+  getFlagFromID(cartItem: CartItem): Flag | undefined {
+    return this.flagService.allFlags.find(flag => (flag.id === cartItem.flagID));
   }
 
   add(id: number): void {
-    this.accountService.addFlagCart(this.currentAccount.username, id)
+    if (this.currentAccount) {
+      this.accountService.addFlagCart(this.currentAccount.username, id)
       .subscribe(() => {
         this.getCart();
       });
+    }
   }
 
   delete(id: number): void {
-    this.accountService.deleteFlagCart(this.currentAccount.username, id)
+    if (this.currentAccount) {
+      this.accountService.deleteFlagCart(this.currentAccount.username, id)
       .subscribe(() => {
         this.getCart();
       });
+    }
   }
 }
